+++ src/index.d.ts
@@ -1,2 +1,2 @@
-export * from './input__value'
-export * from './radio__value'
+export * from './input__value_'
+export * from './radio__value_'