+++ src/input__value_/index.d.ts
@@ -1,6 +1,6 @@
-export declare function input__value(
+export declare function input__value_(
 	selector:string|object, parent?:ParentNode|null
 ):string|undefined
 export {
-	input__value as input_value,
+	input__value_ as input_value,
 }