+++ src/radio__value/index.d.ts
@@ -1,4 +1,0 @@
-export declare function radio__value(name:string):string|undefined
-export {
-	radio__value as radio_value,
-}