+++ src/radio__value_/index.d.ts
@@ -1,0 +1,4 @@
+export declare function radio__value_(name:string):string|undefined
+export {
+	radio__value_ as radio_value,
+}